Add TimeoutManager and related messages
Wrote a TimeoutManager that should be able to receive requests to
monitor a timeout and return if that timeout has expired.

Currently, the TimeoutManager still needs to be able to receive a
stop monitoring message and remove the timeout from the monitored
messages list. However, this shouldn't be too hard. Just a declaration
of a message type and a separate handler for the runTimeoutManager
function.

BullyNode has not been constructed to use the TimeoutManager yet but
integration, again, shouldn't be too difficult (though it will probably
be non-trivial).