Add node discovery, fix bugs, add logging
Added node discovery to bully via the use of the simplelocalnet package.
This mandated an overhaul from the use of processIds to nodeIds for
neighbor discovery. I'm still not quite pleased with this but it works
albeit seemingly in a hacky fashion.

Additionally fixed some bugs that were noticed when I was finally able
to test the bully algorithm itself, the biggest being that nodes were
not notifying the timeoutManager when they received a coordinator
message.

Added log messages using say to help with tracking the process of bully
as it happens within the nodes. This is extremely useful for testing.

One main pain point is my usage of monitor. Right now, if a node rejoins
the network after failure, I start another monitor process to watch it
resulting in 1+n monitoring messages, n being the number of times a node
has come back online. This could be solved by tracking the monitorRefs
in the NodeState and updating them on failure but I have not implemented
this feature yet.